* Major revisit of all admin panel components to use the updated typescript bindings and to pull them exclusively to typescript.

+++ Admin/Source/Page/Create/Create.tsx
@@ -77,7 +77,7 @@ const CreatePage: React.FC = (): React.ReactElement => {
 								name={'description'}
 								label={`Page Description`}
 							/>
-							<TemplateSelector name='pageTemplate' templateType={1} label={`Page Template`} onChange={setSelectedTemplate} />
+							<TemplateSelector name='pageTemplate' templateType={1 as int} label={`Page Template`} onChange={setSelectedTemplate} />
 							<Button type='submit'>{`Create page`}</Button>
 						</Form>
 					</Column>
